pyproject.toml
[project]
name = "multiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
    "google-genai>=1.49.0",
]
